+++ src/Pim/Bundle/CatalogBundle/DependencyInjection/PimCatalogExtension.php
@@ -44,6 +44,7 @@ class PimCatalogExtension extends Extension implements PrependExtensionInterface
         $loader->load('factory.yml');
         $loader->load('entities.yml');
         $loader->load('datagrid_listeners.yml');
+        $loader->load('repositories.yml');
 
         if ($config['record_mails']) {
             $loader->load('mail_recorder.yml');